feat: improving the styles of elements
