+++ index.html
@@ -7,7 +7,12 @@
     <title>Vite App</title>
   </head>
   <body>
-    <div id="app"></div>
+    <div class="navbar" id="header">
+      <span class="title">
+        PDF-Coder
+      </span>
+    </div>
+    <div id="app" />
     <script type="module" src="/src/main.js"></script>
   </body>
 </html>
