+++ src/components/pdf.js
@@ -1,0 +1,32 @@
+import * as pdfjsLib from "pdfjs-dist"
+
+
+//PDF.js worker
+pdfjsLib.GlobalWorkerOptions.workerSrc = '/node_modules/pdfjs-dist/build/pdf.worker.mjs';
+//pdf file render function
+export async function render(file) {
+	const loadingTask = pdfjsLib.getDocument({data: file});
+	const pdf = await loadingTask.promise;
+
+	// Load the first page.
+	const page = await pdf.getPage(1);
+
+	const scale = 1;
+	const viewport = page.getViewport({ scale });
+
+	// Set the canvas dimensions.
+	const canvas = document.getElementById('pdf');
+	const context = canvas.getContext('2d');
+	canvas.height = viewport.height;
+	canvas.width = viewport.width;
+
+	// Render the page into the canvas.
+  if (context === null) return;
+	const renderContext = {
+		canvasContext: context,
+		viewport: viewport,
+	};
+
+  page.render(renderContext);
+	console.log('Page rendered!');
+};
