+++ src/main.js
@@ -1,5 +1,5 @@
 import './style.css'
-import * as pdfjsLib from "pdfjs-dist"
+import { render } from "./components/pdf"
 
 document.querySelector('#app').innerHTML = `
   <div class="app-container">
@@ -9,17 +9,20 @@ document.querySelector('#app').innerHTML = `
       accept="application/pdf,application/msword,
       application/vnd.openxmlformats-officedocument.wordprocessingml.document"
     />
-    <input
+    <button
       data-modal-target="#pdf-modal"
-      type="button"
       id="uploadBtn"
       value="Browse..."
       onclick="document.getElementById('upload').click()"
-    />
-    <div id="pdf-modal">
-      <div data-close-modal id="overlay">
-        <canvas id="pdf"></canvas>
-      </div>
+    >
+      <span id="uploadBtnText">
+        Upload file
+      </span>
+    </button>
+  </div>
+  <div id="pdf-modal">
+    <div data-close-modal id="overlay">
+      <canvas id="pdf"></canvas>
     </div>
   </div>
 `
@@ -61,34 +64,3 @@ function handleFiles(e) {
   if (pdfModal.classList.contains("active")) return;
   pdfModal.classList.add("active");
 }
-
-
-//PDF.js worker
-pdfjsLib.GlobalWorkerOptions.workerSrc = '/node_modules/pdfjs-dist/build/pdf.worker.mjs';
-//pdf file render function
-async function render(file) {
-	const loadingTask = pdfjsLib.getDocument({data: file});
-	const pdf = await loadingTask.promise;
-
-	// Load the first page.
-	const page = await pdf.getPage(1);
-
-	const scale = 1;
-	const viewport = page.getViewport({ scale });
-
-	// Set the canvas dimensions.
-	const canvas = document.getElementById('pdf');
-	const context = canvas.getContext('2d');
-	canvas.height = viewport.height;
-	canvas.width = viewport.width;
-
-	// Render the page into the canvas.
-  if (context === null) return;
-	const renderContext = {
-		canvasContext: context,
-		viewport: viewport,
-	};
-
-  page.render(renderContext);
-	console.log('Page rendered!');
-};
