+++ src/style.css
@@ -25,9 +25,12 @@ a:hover {
 body {
   margin: 0;
   display: flex;
+  flex-direction: column;
   place-items: center;
   min-width: 320px;
   min-height: 100vh;
+  width: 100vw;
+  height: 100vh;
 }
 
 h1 {
@@ -36,12 +39,9 @@ h1 {
 }
 
 #app {
-  max-width: 1280px;
-  margin: 0 auto;
-  padding: 2rem;
   text-align: center;
-  width: 100vh;
-  height: 100vh;
+  width: 100%;
+  height: 100%;
   display: flex;
   flex-direction: column;
 }
@@ -69,17 +69,17 @@ h1 {
 
 button {
   border-radius: 8px;
-  border: 1px solid transparent;
   padding: 0.6em 1.2em;
   font-size: 1em;
   font-weight: 500;
   font-family: inherit;
-  background-color: #1a1a1a;
+  background-color: #88E788;
   cursor: pointer;
-  transition: border-color 0.25s;
+  transition: box-shadow 100ms ease-in-out;
+  color: #5a5c6b;
 }
 button:hover {
-  border-color: #646cff;
+  box-shadow: 0 4px 6px -1px rgb(0 0 0 / 0.1), 0 2px 4px -2px rgb(0 0 0 / 0.1);
 }
 button:focus,
 button:focus-visible {
@@ -99,55 +99,77 @@ button:focus-visible {
   }
 }
 
+canvas {
+  max-width: 100%;
+  max-height: 100%;
+}
+
+.navbar {
+  display: flex;
+  justify-content: start;
+  align-items: center;
+  background-color: #88E788;
+  width: 100%;
+  height: 50px;
+  padding: 6px;
+  box-shadow: 0 10px 15px -3px rgb(0 0 0 / 0.1), 0 4px 6px -4px rgb(0 0 0 / 0.1);
+}
+
+.navbar .title {
+  color: #5a5c6b;
+  font-size: 30px;
+  font-weight: 600;
+  padding: 12px;
+}
+
 .app-container {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    justify-content: start;
-    padding: 3px;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: start;
+  position: relative;
+  width: 100%;
+  height: 100%;
+  padding: 20px;
 }
 
 #upload {
-    display: none;
+  display: none;
 }
 
 #uploadBtn {
-    cursor: pointer;
-    background-color: white;
-    color: #5a5c6b;
-    background-color: #88E788;
-    border-radius: 3px;
-    border-width: 0px;
-    padding-left: 10px;
-    padding-right: 10px;
-    padding-top: 6px;
-    padding-bottom: 6px;
+  border-width: 0px;
+  padding-left: 10px;
+  padding-right: 10px;
+  padding-top: 6px;
+  padding-bottom: 6px;
+  font-size: 20px;
 }
 
 #pdf-modal {
-    position: fixed;
-    top: 50%;
-    left: 50%;
-    transform: translate(-50%, -50%) scale(0);
-    transition: 200ms ease-in-out;
-    width: 100%;
-    height: 100%;
-    display: none;
-    z-index: 10;
-    align-items: center;
-    justify-content: center;
+  position: absolute;
+  top: 50%;
+  left: 50%;
+  transform: translate(-50%, -50%) scale(0);
+  transition: 200ms ease-in-out;
+  width: 100%;
+  height: 100%;
+  display: none;
+  z-index: 10;
+  align-items: center;
+  justify-content: center;
 }
 
 #pdf-modal.active {
-    display: flex;
-    transform: translate(-50%, -50%) scale(1);
+  display: flex;
+  transform: translate(-50%, -50%) scale(1);
 }
 
 #overlay {
-    background-color: rgba(0,0,0,0.25);
-    width: 100%;
-    height: 100%;
-    display: flex;
-    align-items: center;
-    justify-content: center;
+  background-color: rgba(0,0,0,0.25);
+  width: 100%;
+  height: 100%;
+  display: flex;
+  align-items: center;
+  justify-content: center;
 }
